make plots and fix some bugs
The running away bug is still not fixed, but that's not really a big
problem, since it's a 'normal' feature for Ackley. The strange behaviour
around fixed points was fixed by simply pushing down the target some
more in the hopes that the optimizer will continue. This subtly
introduced a division by zero error in the determination of the new
target value which was corrected by avoiding minima that are too close
to the current function value. Success rates are still low on Ackley,
whereas SA gets 100%. Perhaps reading into the source code of SA will
give some insight as to how they proceed with their steps, since that
seems to be the main issue with DM at the moment.