test suite complete, add script to watch test data
Although the test suite is finished, we don't have full reproducibility
because the optimizer does not yet support a way to dump its exact
configuration. That also raises some problems that will be outlined in
the notebook. The test suite also needs to be generalized to simulated
annealing, to facilitate comparisons.