add testing script and switch behaviour of optimizer
Make the optimizer not report failures when all the minima converge to a
point. Instead, make it report success and return the best stuff it has.
It is now STRONGLY discouraged to run the ``conduct_all_experiments''
function directly. Instead, run the ``run_tests.py'' script, which will
test that the repository is clean before trying anything. This will
force you to commit and guarantee reproducibility of tests by being able
to check out a past commit and re-run the exact same experiment and see
what the optimizer configuration was. Planned change: adding the commit
hash to a file in the experiment folder, that way we know what commit
produced this experiment (note that experiments are not put under
version control because they presumably do not change).