Restructure reproducibilit guarantees into own repo.
To facilitate the later distribution of reproducible to others, I've
moved it to its own repo, and then added it as a submodule. This
technique will also be described in the guide that I will write about
using reproducible for data-oriented projects, as well as a kind of git
best practices for bioinformatics projects in general.