WIP rethink callbacks
Inspired by scipy's callbacks for basinhopping, it makes sense for
callbacks to be able to halt the optimization. That make A LOT of sense
since I would want optimization to end when the global minimum is found.
This will allow me to easily make the graph "fraction solved vs
iterations", as I can cause the optimizer to die, by using more
sophisticated callbacks that know what the optimum is.

Making these sophisticated callbacks is trickier: right now the callback
for dm is just a simple function that actually modifies the internal
state of the optimizer. (Of course, the optimizer ignores that
particular state, but it revolts me nonetheless.) So nevermind the
messiness. Plus, that method doesn't generalize to SA since I can't
change SA's internal state while it's running.

The solution is to make callback classes. These classes will implement
__call__, so from the optimizer's point of view, they're just functions,
but in reality, they will maintain in their instances the list of
recorded values. Furthermore, when the callbacks are instantiated, they
can optionally be given the optimum to look for, such that when that
value arises, they can terminate the optimizer. I'll follow in the
footsteps of scipy here: if the callback returns true, then the
optimizer will abort, yielding an OptimizeResult with success set to
True.