rename OptimizeResult -> optimize.Result to satisfy old SciPy
Abacus is using an older version of Anaconda, and thus an older version
of SciPy. Between the two versions, the folks at SciPy changed the API a
bit, moving the class that represents the result of an optimization.
This is what caused my jobs to fail on Abacus all day...