implement the distance ratio distinctness strategy
This strategy allows one to determine if, from the point of view of a
point A, whether two points P and Q are distinct, by requiring that the
distance btween P and Q be at least a constant c times the distance
between A and P.

In the case of the optimizer, we take A to be the iterate, P to be the
current local minimum, and Q to be each past minimum. Of the distinct
minima, we pick the one with the best score as our step direction.