remove termination condition based on step size
If the step size was too small (below the tolerance, currently 10^-4),
then the optimizer would terminate. This is undesirable due to changes
in the steptaking procedure, namely the application of the Strategy
Pattern, and switching to the distance-ratio distinctness strategy.
This strategy is not to guarantee that steps are greater than this
tolerance, so we must remove this termination criterion