refactor contour plotting to do it only once for all figures.
To do so, we also move the nary2binary utility function to dm_utils
module

This function is used for transforming a function of one collection
into a binary function, so that numpy's vectorize function can properly
be applied to it. It used to be a local def inside plot_iterate, but the
refactoring somewhat broke the separation of concerns that existed
there.