implement a simple adaptive steptaking strategy
Adaptive steptaking strategies contrast with the fixed steptaking
strategies by scaling the step vector by a value that depends on the
current state of the optimizer. In this case, we take the percent
difference between the destination cost and the origin cost.

Let Dy be the destination cost, and Sy be the origin cost. We
calculate
    (Sy - Dy) / Sy
to scale the step vector.

This ratio has the property that its values lie in [-1,1]. This is good,
since from experimenting with the fixed step sizes, we found that small
values, close to 0, tended to do better in many problems. Furthermore,
this ratio will guarantee that we always move in the direction of
decreasing cost, which is precisely the goal of the optimizer. If the
direction cost is less than the origin cost, we want the scale to be
positive, to move towards the destination. Otherwise, we want it to be
negative, to move away from the destination.