restructure test functions and implement new ones
The structure of the test functions framework is much clearer, and now
exists in its own module, which can potentially be forked into its own
repository. The dm_tests_config module now deals only with configuring
the optimizers, and the test_functions module deals only with the
different objective functions that we can test the optimizers on.