move plotting function for stepscales out of dm_tests
dm_tests is hell. Instead, I've moved it to the actual plotting script,
rather that have that script import dm_tests just to call one function.

There are also some improvements in the plotter, to deal with missing
data. For instance, it is possible to run this plotting even if one of
DM's or SA's data is missing.

It seems broken, right now.