implement proper fixed distance
Before (and I think this was a flaw in Simon's original code as well),
we offset the function a certain distance rather than actually pick
alternate starting positions. This leads to always choosing essentially
the same starting positions, unless we take care to offset the function
in different directions each time.