hack together a way to use mathematica's FindMinimum from Python
The mathematica_local_min module exports a single function
`mathematica_findminimum` taking keyword arguments serialized to JSON
and passed as stdin to MathematicaFindMinimum.m. They are deserialized
and keywords "point" and "dim" are extracted. The objective function
baked into MathematicaFindMinimum.m is then locally minimum at that
position. The resulting x and y values are serialized to JSON and
printed to stdout. `mathematica_findminimum` captures that output,
deserializes it and returns it.