get rid of evalN Sow-Reap trick for counting function evaluations
In the Before Times, when we didn't know about the EvaluationMonitor
option for NMinimize and FindMinimum, we counted function evaluations by
embedding a Sow call in the objective function, which were then reaped
by the caller. Due to the way the evaluations were counted, this caused
errors when the caller didn't set up the Reap environment correctly
(namely with an additional global variable called evalN).