replace dm_vs_sa_mma_plots with two separate modules
dm_vs_sa_mma_plots was performing statistical calculations as well as
plotting; separation of concerns! Instead, we create
collect_dm_vs_sa_mma.py to get statistics and record them as JSON, which
are then used by a separate plotting module later to create the plots
we're interested in.