remove print statements and fix call to ShiftOverBy
Turns out that the small number of function evaluations is coming from
Simulated Annealing's usual operation. It gives up when it gets stuck
for ~50 iterations. To fix this, we'll implement a random-restart
strategy for simulated annealing, so that it keeps restarting until it
reaches some maximum number of function evaluations (e.g. 20000).